+++ HISTORY.md
@@ -1,4 +1,10 @@
 
+## Next release
+
+### Changed:
+
+- Procedure to evaluate the values of time dimension in case of reduction
+
 ## v1.9.0 - 2024-10-10
 
 ### Fixed: